Add chunked read of multi-contrib contracts
Add a streaming API to read chunks of multi-contrib contracts from the
deploying factory. This means environments that are setup after the
factory has been deployed to retrieve all the deployed contracts from
the chain without needing an archive node.

Previously you would need an archive node and you would need to replay the chain
to locate event logs and reconstruct the list of contracts yourself
which is a lot of boiler plate to get started but also puts the bar very
high.

This can be avoided very simply by recording the contracts as they are deployed.

Also remove some wrapper functions around some public variables and
prefer returning tuples. We are right up against the code-size limit and
so removing these squeezes some bytes out of the contract to give us
breathing room.